requirements.txt
fastapi
uvicorn
pymupdf
chromadb
transformers
torch
python-multipart
# Optional for Gemini LLM
google-generativeai 